+++ internal/provider/byoc_op/byoc_op_projects_resource_store.go
@@ -121,11 +121,26 @@ func (s *byocOpProjectStore) Delete(ctx context.Context, data *BYOCOpProjectReso
 
 	tflog.Info(ctx, fmt.Sprintf("Delete BYOC Op Project request: %+v", request))
 
-	_, err = s.client.DeleteBYOCProject(request)
-	if err != nil {
-		return fmt.Errorf("failed to delete BYOC Op project: %w", err)
+	{
+
+		project, err := s.Describe(ctx, data.ProjectID.ValueString(), data.DataPlaneID.ValueString())
+		if err != nil {
+			return fmt.Errorf("failed to describe BYOC Op project: %w", err)
+		}
+
+		tflog.Info(ctx, fmt.Sprintf("Before delete BYOC Op Project, peek the status: %d", project.Status.ValueInt64()))
+
+		//skip delete if project is already deleted or deleting
+		if !(project.Status.ValueInt64() == int64(BYOCProjectStatusDeleted) || project.Status.ValueInt64() == int64(BYOCProjectStatusDeleting)) {
+			_, err = s.client.DeleteBYOCProject(request)
+			if err != nil {
+				return fmt.Errorf("failed to delete BYOC Op project: %w", err)
+			}
+		}
+
 	}
 
+	// get timeout for delete
 	timeout, diags := data.Timeouts.Delete(ctx, defaultBYOCOpProjectDeleteTimeout)
 	if diags.HasError() {
 		return fmt.Errorf("failed to get delete timeout")
